+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - add --parse option to display the parsed output
 - RdapObject::handle
+- ip address lookup
 
 ### Fixed
 ### Changed